[Added] Filters to select mine sources that should be calculated
It's now possible to select the mine sources for the data analyzis. For
that reason the user must use the command line switch -f followed by the
labels of the data source: E.g: -f source1 source2 sourceN

This commit also enables the total transaction plotting mode, that was
implemented last time.